requirements_to_conda.txt
audioread==3.0.1
biopython==1.79
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.6
dataclassy==1.0.1
decorator==5.1.1
dpcpp-cpp-rt==2024.0.0
fonttools==4.46.0
g2pK==0.9.4
greenlet==3.0.1
idna==3.6
importlib-resources==6.1.1
intel-cmplr-lib-rt==2024.0.0
intel-cmplr-lic-rt==2024.0.0
intel-opencl-rt==2024.0.0
intel-openmp==2024.0.0
jamo==0.4.1
joblib==1.3.2
JPype1==1.4.1
kiwisolver==1.4.5
kneed==0.8.5
konlpy==0.6.0
lazy_loader==0.3
librosa==0.9.2
llvmlite==0.41.1
lxml==4.9.3
markdown-it-py==3.0.0
matplotlib==3.8.2
mdurl==0.1.2
mkl==2024.0.0
mkl-fft==1.3.1
msgpack==1.0.7
nltk==3.8.1
numba==0.58.1
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==4.0.0
pooch==1.8.0
praatio==6.1.0
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-mecab-ko==1.3.3
python-mecab-ko-dic==2.1.1.post2
pytz==2023.3.post1
pyworld==0.3.4
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
resampy==0.4.2
rich==13.7.0
rich-click==1.7.2
scikit-learn==1.2.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.23
tbb==2021.11.0
tgt==1.4.4
threadpoolctl==3.2.0
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
zipp==3.17.0
